Make the starting process message info
When consuming sh in a program, I want to have a verbose setting in
my program that will print what shell commands the program is executing,
but not necessarily the full trace of how that execution is happening.
Putting just the starting command message up to the info level allows it
to still not print by default, and still to print at the debug level.