Cleanup the logging format & make it more useful.
Remove context from the 'command' logger to reduce verbosity & duplication.
Have the child process logger include just the pid as context.

Have the parent print the full command that was executed.

This reduces the length of the log-line (& thus noise) signficantly &
ensures that the entire command is logged (which makes it useful for
debugging).