Actually change directory in pushd context
The pushd context needs to actually change the directory instead of just
setting the _cwd argument. If it doesn't it leads to unexpected behavoir
such as `sh.glob` and `os.getcwd` not working correctly inside of the
pushd context.

This also changes pushd to restore the directory the user was in when
exiting the context, even if the user cd's while inside of the context.
This mirror's bash's pushd/popd behavoir.