Support univeral wheels
sh already supports both py2/py3 from the same code base, so there's no
reason not to produce a universal wheel.

We need to try and import setuptools first (so we can even build wheels
in the first place) and fallback to distutils only if we don't have
setuptools.  (We should prefer setuptools anyways:
https://python-packaging-user-guide.readthedocs.io/en/latest/current/)

Also fix coverage for py3.2.