Remove check for binary data
There's no cheap way we can distinguish between binary data and multibyte strings. It is perfectly valid for decode() to fail since the chunk may be incomplete because of buffering. For example, 'ø' is b'\xc3\xb8' in utf-8, but we could receive the string in two chunks, b'\xc3' and b'\xb8', that each don't decode.

Simply remove the check and trust the user to set the correct out_bufsize. If a fully assembled multibyte line still won't decode, next() will handle it.