Fix for execution contexts from the REPL
On the REPL and from the command line, Python doesn't keep source code
in the stack frame for the toplevel.  Therefore we can't inspect it and
look for the variable being assigned to in order to remove it from
sys.modules.

This is a cornercase but we can hack around it by trying to retrieve the
variable name from the code object instead.  This depends more heavily
on Python internals but this was already code which dove into Python's
internals a bit so perhaps this is okay.

Fixes #422