Raising exceptions should use native strings.
In Python3, the str type is a text string.  On Python2, the str type is
a byte string.  When we raise exceptions we should use native strings
otherwise Python may end up truncating or omitting the exception
message (on Python2) or mangling them (if a byte string were to be given
on Python3).

Fixes #463