Properly raise ErrorReturnCode_0 on exit code 0
This handles the case where exit code 0 is disallowed via '_ok_code' but the
process does exit with this code.

get_exc_from_name() negates signal numbers and calls get_rc_exc() with either
the exit code or the signal number. Exit codes are unsigned ints and thus >=0.
Signal numbers a always >= 1 (see 'kill -l').

get_rc_exc() erroneously handled '0' as a signal when it should be seen
as an exit code.